move components around and capture failed websocket creation
