+++ frontend/src/components/WorkspaceChat/ChatContainer/PromptInput/AgentMenu/index.jsx
@@ -1,12 +1,9 @@
 import { useEffect, useRef, useState } from "react";
 import { Tooltip } from "react-tooltip";
 import { At, Flask, X } from "@phosphor-icons/react";
-import {
-  AGENT_SESSION_END,
-  AGENT_SESSION_START,
-} from "../SlashCommands/endAgentSession";
 import ModalWrapper from "@/components/ModalWrapper";
 import { useModal } from "@/hooks/useModal";
+import { AGENT_SESSION_END, AGENT_SESSION_START } from "@/utils/chat/agent";
 
 export default function AvailableAgentsButton({ showing, setShowAgents }) {
   const [visible, setVisibility] = useState(true);
